refactor(gov): split governance proposal caching
- caching moved to govCache for readability
